Separated RuleSeq module
Moved RuleSeq into a separate file.
